feat(project): refactor ProjectCard import path and add ProjectCard component
