+++ frontend/app/projects/page.tsx
@@ -1,6 +1,6 @@
 import Navigation from '@/components/Navigation';
 import Footer from '@/components/Footer';
-import ProjectCard from '../components/projects/ProjectCard';
+import ProjectCard from '@/components/ProjectCard';
 import { Card } from '@/components/ui/card';
 import { ProjectProps } from '../types/project';
